moving media files to assets
