+++ frontend/src/assets/eloteIcon.png
